refactor(openapi): split 1296-line spec into per-domain modules (#4485)
