+++ .gitignore
@@ -81,3 +81,4 @@ wt-last-updated
 wt-i18n-sessionboard
 node_modules
 aegis
+.openclaw/
